fix(db): separate current from old environments in database
Ref: SRX-7NCIGX
* We move the previous things that happened to the environments to the environments_history table and only keep the current state in the environments table
* The all_environments table is completely removed
* All operations related to environments in database is moved from db.go to a new module db_environments.go
