+++ pkg/db/db.go
@@ -3638,511 +3638,6 @@ func (h *DBHandler) processSingleDeploymentAttemptsRow(ctx context.Context, rows
 
 }
 
-// Environments
-
-type DBAllEnvironments struct {
-	Created      time.Time
-	Version      int64
-	Environments []string
-}
-
-type DBAllEnvironmentsRow struct {
-	Created      time.Time
-	Version      int64
-	Environments string
-}
-
-type DBEnvironment struct {
-	Created      time.Time
-	Name         string
-	Config       config.EnvironmentConfig
-	Applications []string
-}
-
-type DBEnvironmentRow struct {
-	Created      time.Time
-	Name         string
-	Config       string
-	Applications string
-}
-
-func EnvironmentFromRow(_ context.Context, row *DBEnvironmentRow) (*DBEnvironment, error) {
-	//exhaustruct:ignore
-	parsedConfig := config.EnvironmentConfig{}
-	err := json.Unmarshal([]byte(row.Config), &parsedConfig)
-	if err != nil {
-		return nil, fmt.Errorf("unable to unmarshal the JSON in the database, JSON: %s, error: %w", row.Config, err)
-	}
-	applications := []string{}
-	err = json.Unmarshal([]byte(row.Applications), &applications)
-	if err != nil {
-		return nil, fmt.Errorf("unable to unmarshal the JSON in the database, JSON: %s, error: %w", row.Applications, err)
-	}
-	return &DBEnvironment{
-		Created:      row.Created,
-		Name:         row.Name,
-		Config:       parsedConfig,
-		Applications: applications,
-	}, nil
-}
-
-func (h *DBHandler) DBSelectEnvironment(ctx context.Context, tx *sql.Tx, environmentName string) (*DBEnvironment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectEnvironment")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(
-		`
-SELECT created, name, json, applications
-FROM environments
-WHERE name=? AND deleted=false
-ORDER BY version DESC
-LIMIT 1;
-`,
-	)
-	span.SetTag("query", selectQuery)
-	span.SetTag("name", environmentName)
-
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		environmentName,
-	)
-
-	if err != nil {
-		return nil, fmt.Errorf("could not query the environments table for environment %s, error: %w", environmentName, err)
-	}
-	return h.processEnvironmentRow(ctx, rows)
-}
-
-func (h *DBHandler) DBSelectEnvironmentAtTimestamp(ctx context.Context, tx *sql.Tx, environmentName string, ts time.Time) (*DBEnvironment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectEnvironment")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(
-		`
-SELECT created, name, json, applications
-FROM environments
-WHERE name=? AND deleted=false AND created <= ? 
-ORDER BY version DESC
-LIMIT 1;
-`,
-	)
-	span.SetTag("query", selectQuery)
-
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		environmentName,
-		ts,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not query the environments table for environment %s, error: %w", environmentName, err)
-	}
-	return h.processEnvironmentRow(ctx, rows)
-}
-
-func (h *DBHandler) processEnvironmentRow(ctx context.Context, rows *sql.Rows) (*DBEnvironment, error) {
-
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("error while closing row of environments, error: %w", err)
-		}
-	}(rows)
-
-	if rows.Next() {
-		//exhaustruct:ignore
-		row := DBEnvironmentRow{}
-		err := rows.Scan(&row.Created, &row.Name, &row.Config, &row.Applications)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("error scanning the environments table, error: %w", err)
-		}
-		env, err := EnvironmentFromRow(ctx, &row)
-		if err != nil {
-			return nil, err
-		}
-		return env, nil
-	}
-	return nil, nil
-}
-
-func (h *DBHandler) DBSelectEnvironmentsBatch(ctx context.Context, tx *sql.Tx, environmentNames []string) (*[]DBEnvironment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectEnvironmentsBatch")
-	defer span.Finish()
-	if len(environmentNames) > WhereInBatchMax {
-		return nil, fmt.Errorf("SelectEnvironments is not batching queries for now, make sure to not request more than %d environments.", WhereInBatchMax)
-	}
-	if len(environmentNames) == 0 {
-		return &[]DBEnvironment{}, nil
-	}
-
-	selectQuery := h.AdaptQuery(
-		`
-SELECT
-  environments.created AS created,
-  environments.name AS name,
-  environments.json AS json,
-  environments.applications AS applications
-FROM (
-  SELECT
-    MAX(version) AS latest,
-    name
-  FROM
-    environments
-  GROUP BY
-    name) AS latest
-JOIN
-  environments
-ON
-  latest.latest=environments.version
-  AND latest.name = environments.name
-WHERE
-  environments.name IN (?` + strings.Repeat(",?", len(environmentNames)-1) + `)
-  AND deleted=false
-LIMIT ?
-`,
-	)
-	span.SetTag("query", selectQuery)
-	args := []any{}
-	for _, env := range environmentNames {
-		args = append(args, env)
-	}
-	args = append(args, len(environmentNames))
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		args...,
-	)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to query for environments %v, error: %w (query: %s, args: %v)", environmentNames, err, selectQuery, args)
-	}
-
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("error while closing row of environments, error: %w", err)
-		}
-	}(rows)
-
-	envs := []DBEnvironment{}
-	for rows.Next() {
-		//exhaustruct:ignore
-		row := DBEnvironmentRow{}
-		err := rows.Scan(&row.Created, &row.Name, &row.Config, &row.Applications)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("error scanning the environments table, error: %w", err)
-		}
-		env, err := EnvironmentFromRow(ctx, &row)
-		if err != nil {
-			return nil, err
-		}
-		envs = append(envs, *env)
-	}
-	return &envs, nil
-}
-
-// DBWriteEnvironment writes the env to the db
-// if the previousVersion is not supplied, it will do another request to get the environment.
-func (h *DBHandler) DBWriteEnvironment(ctx context.Context, tx *sql.Tx, environmentName string, environmentConfig config.EnvironmentConfig, applications []string) error {
-	span, _ := tracer.StartSpanFromContext(ctx, "DBWriteEnvironment")
-	defer span.Finish()
-
-	if h == nil {
-		return nil
-	}
-	if tx == nil {
-		return fmt.Errorf("attempting to write to the environmets table without a transaction")
-	}
-
-	jsonToInsert, err := json.Marshal(environmentConfig)
-	if err != nil {
-		return fmt.Errorf("error while marshalling the environment config %v, error: %w", environmentConfig, err)
-	}
-	if err != nil {
-		return fmt.Errorf("error while selecting environment %s from database, error: %w", environmentName, err)
-	}
-
-	slices.Sort(applications) // we don't really *need* the sorting, it's just for convenience
-	applicationsJson, err := json.Marshal(applications)
-	if err != nil {
-		return fmt.Errorf("could not marshal the application names list %v, error: %w", applicationsJson, err)
-	}
-
-	insertQuery := h.AdaptQuery(
-		"INSERT Into environments (created, name, json, applications, deleted) VALUES (?, ?, ?, ?, ?);",
-	)
-	now, err := h.DBReadTransactionTimestamp(ctx, tx)
-	if err != nil {
-		return fmt.Errorf("DBWriteEnvironment unable to get transaction timestamp: %w", err)
-	}
-	span.SetTag("query", insertQuery)
-	_, err = tx.Exec(
-		insertQuery,
-		*now,
-		environmentName,
-		jsonToInsert,
-		string(applicationsJson),
-		false,
-	)
-	if err != nil {
-		return fmt.Errorf("could not write environment %s with config %v to environments table, error: %w", environmentName, environmentConfig, err)
-	}
-	return nil
-}
-
-/*
-DBDeleteEnvironment takes the latest config for some env and sets its 'deleted' flag to true. Does not care
-if the applications list is empty or not
-*/
-/*
-DBDeleteEnvironment also updates the overview cache
-*/
-func (h *DBHandler) DBDeleteEnvironment(ctx context.Context, tx *sql.Tx, environmentName string) error {
-	span, _ := tracer.StartSpanFromContext(ctx, "DBDeleteEnvironment")
-	defer span.Finish()
-
-	if h == nil {
-		return nil
-	}
-	if tx == nil {
-		return fmt.Errorf("attempting to write to the environments table without a transaction")
-	}
-
-	currentEnvState, err := h.DBSelectEnvironment(ctx, tx, environmentName)
-	if err != nil {
-		return fmt.Errorf("error while selecting environment %s from database, error: %w", environmentName, err)
-	}
-
-	jsonToInsert, err := json.Marshal(currentEnvState.Config)
-	if err != nil {
-		return fmt.Errorf("error while marshalling the environment config %v, error: %w", currentEnvState.Config, err)
-	}
-
-	slices.Sort(currentEnvState.Applications) // we don't really *need* the sorting, it's just for convenience
-	applicationsJson, err := json.Marshal(currentEnvState.Applications)
-	if err != nil {
-		return fmt.Errorf("could not marshal the application names list %v, error: %w", applicationsJson, err)
-	}
-
-	insertQuery := h.AdaptQuery(
-		"INSERT Into environments (created, name, json, applications, deleted) VALUES (?, ?, ?, ?, ?);",
-	)
-	now, err := h.DBReadTransactionTimestamp(ctx, tx)
-	if err != nil {
-		return fmt.Errorf("DBWriteEnvironment unable to get transaction timestamp: %w", err)
-	}
-	span.SetTag("query", insertQuery)
-	_, err = tx.Exec(
-		insertQuery,
-		*now,
-		environmentName,
-		jsonToInsert,
-		string(applicationsJson),
-		true,
-	)
-	if err != nil {
-		return fmt.Errorf("could not write environment %s with config %v to environments table, error: %w", environmentName, currentEnvState.Config, err)
-	}
-
-	return h.UpdateOverviewDeleteEnvironment(ctx, tx, environmentName)
-}
-
-func (h *DBHandler) UpdateOverviewDeleteEnvironment(ctx context.Context, tx *sql.Tx, environmentName string) error {
-	//Overview cache
-	overview, err := h.ReadLatestOverviewCache(ctx, tx)
-	if overview == nil {
-		//If no overview, there is no need to update it
-		return nil
-	}
-	if err != nil {
-		return fmt.Errorf("unable to read overview cache, error: %w", err)
-	}
-
-	for gIdx, group := range overview.EnvironmentGroups {
-		for idx, currentEnv := range group.Environments {
-			if currentEnv.Name == environmentName {
-				if len(group.Environments) == 1 { //Delete whole group
-					overview.EnvironmentGroups = append(overview.EnvironmentGroups[:gIdx], overview.EnvironmentGroups[gIdx+1:]...)
-				} else {
-					overview.EnvironmentGroups[gIdx].Environments = append(group.Environments[:idx], group.Environments[idx+1:]...)
-				}
-				break
-			}
-		}
-	}
-	err = h.WriteOverviewCache(ctx, tx, overview)
-	if err != nil {
-		return fmt.Errorf("Unable to write overview cache, error: %w", err)
-	}
-	return nil
-}
-
-func (h *DBHandler) DBSelectAllEnvironments(ctx context.Context, transaction *sql.Tx) (*DBAllEnvironments, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllEnvironments")
-	defer span.Finish()
-
-	if h == nil {
-		return nil, nil
-	}
-	if transaction == nil {
-		return nil, fmt.Errorf("no transaction provided when selecting all environments from all_environments table")
-	}
-
-	selectQuery := h.AdaptQuery(
-		"SELECT created, version, json FROM all_environments ORDER BY version DESC LIMIT 1;",
-	)
-
-	rows, err := transaction.QueryContext(ctx, selectQuery)
-	if err != nil {
-		return nil, fmt.Errorf("error while execuring query to get all environments, error: %w", err)
-	}
-
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("error while closing row on all_environments table, error: %w", err)
-		}
-	}(rows)
-
-	if rows.Next() {
-		//exhaustruct:ignore
-		row := DBAllEnvironmentsRow{}
-
-		err := rows.Scan(&row.Created, &row.Version, &row.Environments)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("error while scanning all_environments row, error: %w", err)
-		}
-
-		parsedEnvironments := make([]string, 0)
-		err = json.Unmarshal([]byte(row.Environments), &parsedEnvironments)
-		if err != nil {
-			return nil, fmt.Errorf("error occured during JSON unmarshalling, JSON: %s, error: %w", row.Environments, err)
-		}
-
-		err = closeRows(rows)
-		if err != nil {
-			return nil, fmt.Errorf("error while closing rows, error: %w", err)
-		}
-
-		return &DBAllEnvironments{
-			Created:      row.Created,
-			Version:      row.Version,
-			Environments: parsedEnvironments,
-		}, nil
-	}
-
-	err = closeRows(rows)
-	if err != nil {
-		return nil, fmt.Errorf("error while closing rows, error: %w", err)
-	}
-	return nil, nil
-}
-
-func (h *DBHandler) DBWriteAllEnvironments(ctx context.Context, transaction *sql.Tx, environmentNames []string) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBWriteAllEnvironments")
-	defer span.Finish()
-
-	slices.Sort(environmentNames) // we don't really *need* the sorting, it's just for convenience
-	jsonToInsert, err := json.Marshal(environmentNames)
-	if err != nil {
-		return fmt.Errorf("could not marshal the environment names list %v, error: %w", environmentNames, err)
-	}
-
-	allEnvironments, err := h.DBSelectAllEnvironments(ctx, transaction)
-	if err != nil {
-		return fmt.Errorf("unable to select all environments, error: %w", err)
-	}
-
-	previousVersion := int64(0)
-	if allEnvironments != nil {
-		previousVersion = allEnvironments.Version
-	}
-
-	insertQuery := h.AdaptQuery("INSERT INTO all_environments (created, version, json) VALUES (?, ?, ?)")
-	now, err := h.DBReadTransactionTimestamp(ctx, transaction)
-	if err != nil {
-		return fmt.Errorf("DBWriteAllEnvironments unable to get transaction timestamp: %w", err)
-	}
-	span.SetTag("query", insertQuery)
-	_, err = transaction.Exec(
-		insertQuery,
-		*now,
-		previousVersion+1,
-		jsonToInsert,
-	)
-	if err != nil {
-		return fmt.Errorf("unable to perform the insert query, error: %w", err)
-	}
-
-	return nil
-}
-
-func (h *DBHandler) DBSelectAnyEnvironment(ctx context.Context, tx *sql.Tx) (*DBAllEnvironments, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAnyEnvironment")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(
-		"SELECT created, version, json FROM all_environments ORDER BY version DESC LIMIT 1;",
-	)
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(ctx, selectQuery)
-	if err != nil {
-		return nil, fmt.Errorf("could not query the all_environments table, error: %w", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("error while closing the row of all_environments, error: %w", err)
-		}
-	}(rows)
-
-	//exhaustruct:ignore
-	row := DBAllEnvironmentsRow{}
-	if rows.Next() {
-		err := rows.Scan(&row.Created, &row.Version, &row.Environments)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("error scanning the results of the query for selecting any row in all_environments, error: %w", err)
-		}
-		err = closeRows(rows)
-		if err != nil {
-			return nil, fmt.Errorf("error while closing the rows of the query for selecting any row in all_environments, error: %w", err)
-		}
-
-		jsonData := make([]string, 0)
-		err = json.Unmarshal([]byte(row.Environments), &jsonData)
-
-		if err != nil {
-			return nil, fmt.Errorf("error parsing the value of the JSON column of the all_environments table, JSON content: %s, error: %w", row.Environments, err)
-		}
-
-		return &DBAllEnvironments{
-			Version:      row.Version,
-			Created:      row.Created,
-			Environments: jsonData,
-		}, nil
-	}
-
-	err = closeRows(rows)
-	if err != nil {
-		return nil, fmt.Errorf("error while closing the rows of the query for selecting any row in all_environments (where no rows were returned), error: %w", err)
-	}
-
-	return nil, nil
-}
-
 func (h *DBHandler) RunCustomMigrationEnvironments(ctx context.Context, getAllEnvironmentsFun GetAllEnvironmentsFun) error {
 	span, ctx := tracer.StartSpanFromContext(ctx, "RunCustomMigrationEnvironments")
 	defer span.Finish()
@@ -4160,13 +3655,11 @@ func (h *DBHandler) RunCustomMigrationEnvironments(ctx context.Context, getAllEn
 			return fmt.Errorf("could not get environments, error: %w", err)
 		}
 
-		allEnvironmentNames := make([]string, 0)
 		allEnvsApps, err := h.FindEnvsAppsFromReleases(ctx, transaction)
 		if err != nil {
 			return err
 		}
 		for envName, config := range allEnvironments {
-			allEnvironmentNames = append(allEnvironmentNames, envName)
 			if allEnvsApps[envName] == nil {
 				allEnvsApps[envName] = make([]string, 0)
 			}
@@ -4174,10 +3667,6 @@ func (h *DBHandler) RunCustomMigrationEnvironments(ctx context.Context, getAllEn
 			if err != nil {
 				return fmt.Errorf("unable to write manifest for environment %s to the database, error: %w", envName, err)
 			}
-		}
-		err = h.DBWriteAllEnvironments(ctx, transaction, allEnvironmentNames)
-		if err != nil {
-			return fmt.Errorf("unable to write to write all environments list to the database, error: %w", err)
 		}
 		return nil
 	})
@@ -4208,7 +3697,7 @@ func (h *DBHandler) RunCustomMigrationEnvironmentApplications(ctx context.Contex
 			return fmt.Errorf("could not get environments, error: %w", err)
 		}
 		var allEnvsApps map[string][]string
-		for _, envName := range allEnvironments.Environments {
+		for _, envName := range allEnvironments {
 			env, err := h.DBSelectEnvironment(ctx, transaction, envName)
 			if err != nil {
 				return fmt.Errorf("could not get env: %s, error: %w", envName, err)
