+++ pkg/db/overview.go
@@ -114,6 +114,36 @@ func (h *DBHandler) UpdateOverviewEnvironmentLock(ctx context.Context, transacti
 	return nil
 }
 
+func (h *DBHandler) UpdateOverviewDeleteEnvironment(ctx context.Context, tx *sql.Tx, environmentName string) error {
+	//Overview cache
+	overview, err := h.ReadLatestOverviewCache(ctx, tx)
+	if overview == nil {
+		//If no overview, there is no need to update it
+		return nil
+	}
+	if err != nil {
+		return fmt.Errorf("unable to read overview cache, error: %w", err)
+	}
+
+	for gIdx, group := range overview.EnvironmentGroups {
+		for idx, currentEnv := range group.Environments {
+			if currentEnv.Name == environmentName {
+				if len(group.Environments) == 1 { //Delete whole group
+					overview.EnvironmentGroups = append(overview.EnvironmentGroups[:gIdx], overview.EnvironmentGroups[gIdx+1:]...)
+				} else {
+					overview.EnvironmentGroups[gIdx].Environments = append(group.Environments[:idx], group.Environments[idx+1:]...)
+				}
+				break
+			}
+		}
+	}
+	err = h.WriteOverviewCache(ctx, tx, overview)
+	if err != nil {
+		return fmt.Errorf("Unable to write overview cache, error: %w", err)
+	}
+	return nil
+}
+
 func (h *DBHandler) IsOverviewEmpty(overviewResp *api.GetOverviewResponse) bool {
 	if overviewResp == nil {
 		return true
